requirements.txt
faiss-cpu
sentence-transformers
transformers
torch
PyPDF2
tqdm
requests
fastapi
uvicorn[standard]
python-multipart
